rename CollectionListPage to Collection, CollectionList to ItemList, add LibraryHome for user home page
